Move excon_defaults to SharedNativeHelpers
