+++ bundler/helpers/lib/functions/version_resolver.rb
@@ -1,10 +1,8 @@
+require "shared_native_helpers"
+
 module Functions
   class VersionResolver
     GEM_NOT_FOUND_ERROR_REGEX = /locked to (?<name>[^\s]+) \(/.freeze
-    USER_AGENT = "dependabot-core/bundler-helper "\
-                 "#{Excon::USER_AGENT} ruby/#{RUBY_VERSION} "\
-                 "(#{RUBY_PLATFORM}) "\
-                "(+https://github.com/dependabot/dependabot-core)"
 
     attr_reader :dependency_name, :dependency_requirements,
                 :gemfile_name, :lockfile_name
@@ -140,7 +138,7 @@ module Functions
       versions = Excon.get(
         "#{fetcher.fetch_uri}api/v1/versions/#{dependency_name}.json",
         idempotent: true,
-        **excon_defaults
+        **SharedNativeHelpers.excon_defaults
       )
 
       # Give the benefit of the doubt if something goes wrong fetching
@@ -169,32 +167,6 @@ module Functions
       return nil unless gemfile_name
 
       @ruby_version ||= build_definition([]).ruby_version&.gem_version
-    end
-
-    def excon_middleware
-      Excon.defaults[:middlewares] +
-        [Excon::Middleware::Decompress] +
-        [Excon::Middleware::RedirectFollower]
-    end
-
-    def excon_headers(headers = nil)
-      headers ||= {}
-      {
-        "User-Agent" => USER_AGENT
-      }.merge(headers)
-    end
-
-    def excon_defaults(options = nil)
-      options ||= {}
-      headers = options.delete(:headers)
-      {
-        connect_timeout: 5,
-        write_timeout: 5,
-        read_timeout: 20,
-        omit_default_port: true,
-        middlewares: excon_middleware,
-        headers: excon_headers(headers)
-      }.merge(options)
     end
   end
 end
